Modifed model exporting. Created separate route for workouts. Added password field to users to users.
